+++ packages/components/editor/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/editor
 
+## 0.0.41
+
+### Patch Changes
+
+- added yt to editor
+
 ## 0.0.40
 
 ### Patch Changes